+++ crates/litho-extract/src/extractors/python.rs
@@ -31,7 +31,7 @@ fn identifier_child(node: tree_sitter::Node<'_>, source_bytes: &[u8]) -> Option<
         }
     }
     for i in 0..node.child_count() {
-        if let Some(child) = node.child(i as u32)
+        if let Some(child) = node.child(i as usize)
             && child.kind() == "identifier"
         {
             let t = node_text(child, source_bytes);
@@ -73,7 +73,7 @@ impl Extractor for PythonExtractor {
                     // Recurse into decorated definitions to find the actual def.
                     let child_count = node.child_count();
                     for i in (0..child_count).rev() {
-                        if let Some(child) = node.child(i as u32) {
+                        if let Some(child) = node.child(i as usize) {
                             stack.push(child);
                         }
                     }
@@ -105,14 +105,14 @@ impl Extractor for PythonExtractor {
                     kind: sk.to_string(),
                     visibility: vis.to_string(),
                     signature,
-                    line: node.start_position().row + 1,
+                    line: node.start_position().row as usize + 1,
                 });
                 continue;
             }
 
             let child_count = node.child_count();
             for i in (0..child_count).rev() {
-                if let Some(child) = node.child(i as u32) {
+                if let Some(child) = node.child(i as usize) {
                     stack.push(child);
                 }
             }
@@ -176,7 +176,7 @@ impl Extractor for PythonExtractor {
 
             let child_count = node.child_count();
             for i in (0..child_count).rev() {
-                if let Some(child) = node.child(i as u32) {
+                if let Some(child) = node.child(i as usize) {
                     stack.push(child);
                 }
             }
